requirements.txt
crewai
python-telegram-bot
requests
